requirements.txt
discord.py
youtube_dl
PyNaCl
aiohttp